+++ hata/discord/client/client.py
@@ -3740,7 +3740,7 @@ class Client(ClientUserPBase):
         
         Returns
         -------
-        message : ``Message``, `None`
+        message : `None`, ``Message``
             Returns `None` if there is nothing to send.
         
         Raises
@@ -9599,7 +9599,8 @@ class Client(ClientUserPBase):
         return thread_channels
     
     
-    async def thread_create(self, message_or_channel, name, *, auto_archive_after=None, type_=None, invitable=True):
+    async def thread_create(self, message_or_channel, name, *, auto_archive_after=None, type_=None, invitable=True,
+            slowmode = None):
         """
         Creates a new thread derived from the given message or channel.
         
@@ -9620,9 +9621,6 @@ class Client(ClientUserPBase):
         
         auto_archive_after : `None`, `int` = `None`, Optional (Keyword only)
             The duration in seconds after the thread auto archives. Can be any of `3600`, `86400`, `259200`, `604800`.
-            
-            > For `259200` seconds (or 3 days) or `604800` seconds (or 7 days) auto archive duration the guild needs
-            > to be boosted.
         
         type_ : `None`, `int` = `None`, Optional (Keyword only)
             The thread channel's type to create. Can be either `10`,`11`,`12`.
@@ -9632,9 +9630,13 @@ class Client(ClientUserPBase):
             
             Applicable for private threads. Defaults to `True`.
         
+        slowmode : `int`, Optional (Keyword only)
+            Slowmode for the new thread channel.
+        
         Returns
         -------
         thread_channel : ``Channel``
+            The created thread channel.
         
         Raises
         ------
@@ -9670,32 +9672,27 @@ class Client(ClientUserPBase):
             message_id = None
             channel = message_or_channel
             channel_id = channel.id
-            guild = channel.guild
         
         elif isinstance(message_or_channel, Message):
             message_id = message_or_channel.id
             channel_id = message_or_channel.channel_id
             channel = CHANNELS.get(channel_id, None)
-            guild = channel.guild
         
         else:
             channel_id = maybe_snowflake(message_or_channel)
             if (channel_id is not None):
                 message_id = None
                 channel = CHANNELS.get(channel_id, None)
-                guild = None
             
             elif isinstance(message_or_channel, MessageRepr):
                 message_id = message_or_channel.id
                 channel_id = message_or_channel.channel_id
                 channel = CHANNELS.get(channel_id, None)
-                guild = message_or_channel.guild
             
             elif isinstance(message_or_channel, MessageReference):
                 message_id = message_or_channel.message_id
                 channel_id = message_or_channel.channel_id
                 channel = CHANNELS.get(channel_id, None)
-                guild = message_or_channel.guild
             
             else:
                 snowflake_pair = maybe_snowflake_pair(message_or_channel)
@@ -9708,7 +9705,6 @@ class Client(ClientUserPBase):
                 
                 channel_id, message_id = snowflake_pair
                 channel = CHANNELS.get(channel_id)
-                guild = None
         
         if __debug__:
             if not isinstance(name, str):
@@ -9758,6 +9754,9 @@ class Client(ClientUserPBase):
             'type': type_,
         }
         
+        _maybe_add_channel_slowmode_field_to_data(type_, None, data, slowmode)
+        
+        
         if (type_ == CHANNEL_TYPES.guild_thread_private) and (not invitable):
             data['invitable'] = invitable
         
@@ -9766,17 +9765,279 @@ class Client(ClientUserPBase):
             coroutine = self.http.thread_create(channel_id, data)
         else:
             coroutine = self.http.thread_create_from_message(channel_id, message_id, data)
-        
         channel_data = await coroutine
-        if guild is None:
-            guild_id = channel_data.get('guild_id', None)
-            if (guild_id is None):
-                guild = None
-            else:
-                guild_id = int(guild_id)
-                guild = GUILDS.get(guild_id, None)
         
-        return Channel(channel_data, self, guild.id)
+        if channel is None:
+            guild_id = channel_data.get('guild_id', None)
+            if (guild_id is not None):
+                guild_id = int(guild_id)
+        else:
+            guild_id = channel.guild_id
+        
+        return Channel(channel_data, guild_id)
+    
+    
+    async def forum_thread_create(
+        self, channel_forum, name, content=None, *, allowed_mentions=...,  auto_archive_after=None,
+        components=None, embed=None, file=None, nonce=None, reply_fail_fallback=False, slowmode=None, sticker=None,
+        suppress_embeds=False, tts=False
+    ):
+        """
+        Creates and thread at the given `channel` with the given message fields. If there is nothing to send will
+        return `None` for channel and `None` for message as well.
+        
+        This method is a coroutine.
+        
+        Parameters
+        ----------
+        channel_forum : ``Channel``, `int`
+            The forum channel's identifier where the thread will be started.
+        
+        name : `str`
+            The created thread's name.
+        
+        content : `None`, `str`, ``EmbedBase``, `Any` = `None`, Optional
+            The message's content if given. If given as `str` or empty string, then no content will be sent, meanwhile
+            if any other non `str`, ``EmbedBase`` is given, then will be casted to string.
+            
+            If given as ``EmbedBase``, then is sent as the message's embed.
+        
+        allowed_mentions : `None`,  `str`, ``UserBase``, ``Role``, `list` of (`str`, ``UserBase``, ``Role`` )
+                , Optional (Keyword only)
+            Which user or role can the message ping (or everyone). Check ``parse_allowed_mentions`` for details.
+        
+        auto_archive_after : `None`, `int` = `None`, Optional (Keyword only)
+            The duration in seconds after the thread auto archives. Can be any of `3600`, `86400`, `259200`, `604800`.
+        
+        components : `None`, ``ComponentBase``, (`tuple`, `list`) of (``ComponentBase``, (`tuple`, `list`) of
+                ``ComponentBase``) = `None`, Optional (Keyword only)
+            Components attached to the message.
+            
+            > `components` do not count towards having any content in the message.
+        
+        embed : ``EmbedBase``, `list` of ``EmbedBase`` = `None`, Optional (Keyword only)
+            The embedded content of the message.
+            
+            If `embed` and `content` parameters are both given as  ``EmbedBase``, then `TypeError` is raised.
+        
+        file : `None`, `Any` = `None`, Optional (Keyword only)
+            A file or files to send. Check ``create_file_form`` for details.
+        
+        reply_fail_fallback : `bool` = `False`, Optional (Keyword only)
+            Whether normal message should be sent if the referenced message is deleted. Defaults to `False`.
+        
+        nonce : `None`, `str` = `None`, Optional (Keyword only)
+            Used for optimistic message sending. Will shop up at the message's data.
+        
+        slowmode : `int`, Optional (Keyword only)
+            Slowmode for the new thread channel.
+        
+        sticker : `None`, ``Sticker``, `int`, (`list`, `set`, `tuple`) of (``Sticker``, `int`) = `None` \
+                , Optional (Keyword only)
+            Sticker or stickers to send within the message.
+        
+        suppress_embeds : `bool` = `False`, Optional (Keyword only)
+            Whether the message's embeds should be suppressed initially.
+        
+        tts : `bool` = `False`, Optional (Keyword only)
+            Whether the message is text-to-speech.
+        
+        Returns
+        -------
+        thread_channel : `None`, ``Channel``
+            The created thread channel. `None` if there was nothing to send.
+        message : `None`, ``Message``
+            Returns `None` if there is nothing to send.
+        
+        Raises
+        ------
+        TypeError
+            - If a parameter's type is incorrect.
+        ValueError
+            - If a parameter's value is incorrect.
+        ConnectionError
+            No internet connection.
+        DiscordException
+            If any exception was received from the Discord API.
+        
+        See Also
+        --------
+        ``.message_create`` : Sending a message to a text channel.
+        ``.thread_create`` : Create thread in a text channel.
+        """
+        channel, channel_id = get_channel_and_id(channel_forum, Channel.is_guild_forum)
+        
+        # name & auto_archive_after
+        
+        if __debug__:
+            if not isinstance(name, str):
+                raise AssertionError(
+                    f'`name` can be `str`, got {name.__class__.__name__}; {name!r}.'
+                )
+            
+            name_length = len(name)
+            
+            if (name_length) < 2 or (name_length > 100):
+                raise AssertionError(
+                    f'`name` length can be in range [2:100], got {name_length}; {name!r}.'
+                )
+        
+        if auto_archive_after is None:
+            if channel is None:
+                auto_archive_after = AUTO_ARCHIVE_DEFAULT
+            else:
+                auto_archive_after = channel.default_auto_archive_after
+        else:
+            if __debug__:
+                if not isinstance(auto_archive_after, int):
+                    raise AssertionError(
+                        f'`auto_archive_after` can be `None`, `datetime`, got '
+                        f'{auto_archive_after.__class__.__name__}; {auto_archive_after!r}.'
+                    )
+                
+                if auto_archive_after not in AUTO_ARCHIVE_OPTIONS:
+                    raise AssertionError(
+                        f'`auto_archive_after` can be any of: {AUTO_ARCHIVE_OPTIONS}, got {auto_archive_after!r}.'
+                    )
+        
+        content, embed = validate_content_and_embed(content, embed, False)
+        
+        # Sticker check order:
+        # 1.: None -> None
+        # 2.: Sticker -> [sticker.id]
+        # 3.: int (str) -> [sticker]
+        # 4.: (list, tuple) of (Sticker, int (str)) -> [sticker.id / sticker, ...] / None
+        # 5.: raise
+        
+        if sticker is None:
+            sticker_ids = None
+        else:
+            sticker_ids = []
+            if isinstance(sticker, Sticker):
+                sticker_id = sticker.id
+                sticker_ids.append(sticker_id)
+            else:
+                sticker_id = maybe_snowflake(sticker)
+                if sticker_id is None:
+                    if isinstance(sticker, (list, tuple)):
+                        for sticker_element in sticker:
+                            if isinstance(sticker_element, Sticker):
+                                sticker_id = sticker_element.id
+                            else:
+                                sticker_id = maybe_snowflake(sticker_element)
+                                if sticker_id is None:
+                                    raise TypeError(
+                                        f'`sticker` can contain only `{Sticker.__name__}`, `int` elements, '
+                                        f'got {sticker_element.__class__.__name__}; {sticker_element!r}; '
+                                        f'sticker={sticker!r}.'
+                                    )
+                            
+                            sticker_ids.append(sticker_id)
+                        
+                        if not sticker_ids:
+                            sticker_ids = None
+                    else:
+                        raise TypeError(
+                            f'`sticker` can be `None`, `{Sticker.__name__}`, `int`, '
+                            f'(`list`, `tuple`) of (`{Sticker.__name__}`, `int`), got '
+                            f'{sticker.__class__.__name__}; {sticker!r}.'
+                        )
+                else:
+                    sticker_ids.append(sticker_id)
+        
+        components = get_components_data(components, False)
+        
+        if __debug__:
+            
+            if (nonce is not None) and (not isinstance(nonce, str)):
+                raise AssertionError(
+                    f'`nonce` can be `None`, `str`, got {nonce.__class__.__name__}; {nonce!r}.'
+                )
+            
+            if not isinstance(reply_fail_fallback, bool):
+                raise AssertionError(
+                    f'`reply_fail_fallback` can be `bool`, got {reply_fail_fallback.__class__.__name__}; '
+                    f'{reply_fail_fallback!r}.')
+            
+            if not isinstance(suppress_embeds, bool):
+                raise AssertionError(
+                    f'`suppress_embeds` can be `bool`, got {suppress_embeds.__class__.__name__}; {suppress_embeds!r}.'
+                )
+        
+            if not isinstance(tts, bool):
+                raise AssertionError(
+                    f'`tts` can be `bool`, got {tts.__class__.__name__}, {tts!r}.'
+                )
+        
+        # Build payload
+        message_data = {}
+        contains_content = False
+        
+        if (content is not None):
+            message_data['content'] = content
+            contains_content = True
+        
+        if (embed is not None):
+            message_data['embeds'] = [embed.to_data() for embed in embed]
+            contains_content = True
+        
+        if (sticker_ids is not None):
+            message_data['sticker_ids'] = sticker_ids
+            contains_content = True
+        
+        if (components is not None):
+            message_data['components'] = components
+        
+        if tts:
+            message_data['tts'] = True
+        
+        if (nonce is not None):
+            message_data['nonce'] = nonce
+        
+        if (allowed_mentions is not ...):
+            message_data['allowed_mentions'] = parse_allowed_mentions(allowed_mentions)
+        
+        if (message_id is not None):
+            message_reference_data = {'message_id': message_id}
+            
+            if reply_fail_fallback:
+                message_reference_data['fail_if_not_exists'] = False
+            
+            message_data['message_reference'] = message_reference_data
+        
+        if suppress_embeds:
+            message_data['flags'] = MESSAGE_FLAG_VALUE_SUPPRESS_EMBEDS
+        
+        message_data = add_file_to_message_data(message_data, file, contains_content, False)
+        if message_data is None:
+            return None, None
+        
+        data = {
+            'name': name,
+            'auto_archive_duration': auto_archive_after // 60,
+            'message': message_data,
+        }
+        
+        _maybe_add_channel_slowmode_field_to_data(CHANNEL_TYPES.guild_thread_public, None, data, slowmode)
+        
+        channel_data = await self.http.thread_create(channel_id, data)
+        
+        if channel is None:
+            guild_id = channel_data.get('guild_id', None)
+            if (guild_id is not None):
+                guild_id = int(guild_id)
+        else:
+            guild_id = channel.guild_id
+        
+        thread_channel = Channel(channel_data, self, guild_id)
+        
+        message_data = thread_channel.get('message', None)
+        if (message_data is None):
+            message = None
+        else:
+            message = thread_channel._create_new_message(message_data)
+        
+        return thread_channel, message
     
     
     async def thread_join(self, thread_channel):